Refactor: set PictureView toolbar and system bar colors on runtime
Using flags instead of styles + XML attributes allows the color to be anything, translucent or otherwise.
Don't really know if setting custom toolbar (instead of using default actionbar) is required for this method to work.
Only downside is that toolbar has to be manually offset by the status bar height to prevent overlap (setting fitsSystemWindow=true prevents draw-under even inside CoordinatorLayout for whatever reason).
Status and toolbar also have exactly the same color now, as a side-effect of the changes.