Fix: misnomer deep zoom -> true zoom; gifs in true zoom
Deep zoom is a misleading name because if the image is small to start with,
you'd actually be able to zoom in *less* compared to in the default mode.

True zoom is probably a more accurate name: zoom level (while maintaining clarity)
for small images is small, and for large images is large.

Also fixed animated images not showing in true zoom: gifs will play, other types
will show still image b/c of Glide's particularities.